Loan Approval Prediction System/requirements.txt
# Web Framework
Flask>=3.0.0
Werkzeug>=3.0.0
flask-cors>=4.0.0

# Machine Learning
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
xgboost>=2.0.0
joblib>=1.3.0

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

Flask-Login>=0.6.0
Flask-SQLAlchemy>=3.0.0
Flask-WTF>=1.2.0
email-validator>=2.0.0
Werkzeug>=3.0.0  # For password hashing